+++ db/migrate/20180610024859_create_invoices.rb
@@ -3,9 +3,9 @@ class CreateInvoices < ActiveRecord::Migration[5.2]
     create_table :invoices do |t|
       t.references :client, foreign_key: true
       t.references :specialist, foreign_key: true
-      t.decimal :sub_total, precision: 5, scale: 2
-      t.decimal :total, precision: 5, scale: 2
-      t.decimal :iva, precision: 5, scale: 2
+      t.float :sub_total, precision: 5, scale: 2
+      t.float :total, precision: 5, scale: 2
+      t.float :iva, precision: 5, scale: 2
 
       # t.actable
       t.timestamps
@@ -16,9 +16,9 @@ class CreateInvoices < ActiveRecord::Migration[5.2]
       t.references :supplier, foreign_key: true
       # t.string     :number
 
-      t.decimal :sub_total, precision: 5, scale: 2
-      t.decimal :total, precision: 5, scale: 2
-      t.decimal :iva, precision: 5, scale: 2
+      t.float :sub_total, precision: 5, scale: 2
+      t.float :total, precision: 5, scale: 2
+      t.float :iva, precision: 5, scale: 2
 
       t.timestamps
     end
@@ -27,9 +27,9 @@ class CreateInvoices < ActiveRecord::Migration[5.2]
       t.references :income_product, foreign_key: true
       # t.references :item, polymorphic: true, index: true
       t.references :product, foreign_key: true
-      t.decimal :quantity
-      t.decimal :price_unit, precision: 5, scale: 2
-      t.decimal :price_total, precision: 5, scale: 2
+      t.float :quantity
+      t.float :price_unit, precision: 5, scale: 2
+      t.float :price_total, precision: 5, scale: 2
 
     end
