+++ db/migrate/20180610025921_create_invoice_details.rb
@@ -4,9 +4,9 @@ class CreateInvoiceDetails < ActiveRecord::Migration[5.2]
       t.references :invoice, foreign_key: true
       # t.references :item, polymorphic: true, index: true
       t.references :item, foreign_key: true
-      t.decimal :quantity
-      t.decimal :price_unit, precision: 5, scale: 2
-      t.decimal :price_total, precision: 5, scale: 2
+      t.float :quantity
+      t.float :price_unit, precision: 5, scale: 2
+      t.float :price_total, precision: 5, scale: 2
 
     end
   end
